twitter_topic_mobility_estimation/requirements.txt
mecab-python3
PyMySql
sqlalchemy
numpy
gensim
scikit-learn
pandas
plotly
geopy
matplotlib